'Refactored by Sourcery'

+++ src/drozer/modules/common/file_system.py
@@ -23,26 +23,19 @@ class FileSystem(object):
 
         file_io = self.new("java.io.File", source)
 
-        if file_io.exists() == True:
-            return file_io.delete()
-        else:
-            return None
+        return file_io.delete() if file_io.exists() == True else None
 
     def downloadFile(self, source, destination, block_size=65536):
         """
         Copy a file from the Agent's file system to the local one.
         """
 
-        data = self.readFile(source, block_size=block_size)
-
-        if data:
+        if data := self.readFile(source, block_size=block_size):
             if os.path.isdir(destination):
                 destination = os.path.sep.join([destination, source.split("/")[-1]])
-                
-            output = open(destination, 'wb')
-            output.write(str(data))
-            output.close()
 
+            with open(destination, 'wb') as output:
+                output.write(str(data))
             return len(data)
         else:
             return None
@@ -76,25 +69,18 @@ class FileSystem(object):
 
         file_io = self.new("java.io.File", source)
 
-        if file_io.exists() == True:
-            return file_io.length()
-        else:
-            return None
+        return file_io.length() if file_io.exists() == True else None
     
     def format_file_size(self, size):
         """
         Return the size of a file in human-readable form (i.e., x KiB).
         """
-        
+
         for x in ['bytes', 'KiB', 'MiB', 'GiB']:
             if size < 1024.0 and size > -1024.0:
-                if x != "bytes":
-                    return "%.1f %s" % (size, x)
-                else:
-                    return "%d %s" % (size, x)
-            
+                return "%.1f %s" % (size, x) if x != "bytes" else "%d %s" % (size, x)
             size /= 1024.0
-            
+
         return "%3.1f%s" % (size, 'TiB')
 
     def isDirectory(self, target):
@@ -121,8 +107,8 @@ class FileSystem(object):
         """
         #TODO does not work past the first folder
         file_io = self.new("java.io.File", target)
-        
-        return ["%s%s" %(s, '/') if file_io.isDirectory() else s for s in file_io.list()]
+
+        return [f"{s}/" if file_io.isDirectory() else s for s in file_io.list()]
         
     def md5sum(self, source):
         """
@@ -133,10 +119,7 @@ class FileSystem(object):
 
         file_io = self.new("java.io.File", source)
 
-        if file_io.exists() == True:
-            return FileUtil.md5sum(file_io)
-        else:
-            return None
+        return FileUtil.md5sum(file_io) if file_io.exists() == True else None
 
     def readFile(self, source, block_size=65536):
         """
@@ -151,10 +134,10 @@ class FileSystem(object):
             file_stream = self.new("java.io.FileInputStream", file_io)
 
             data = ""
-            
+
             while True:
                 block = ByteStreamReader.read(file_stream, 0, block_size).base64_encode()
-                
+
                 if len(block) > 0:
                     data += base64.decodestring(block)
                 else:
@@ -195,7 +178,7 @@ class FileSystem(object):
                 ByteStreamWriter.writeHexStream(file_stream, binascii.hexlify(c))
 
             file_stream.close()
-            
+
             return len(data)
         else:
             return None
